Fix a bug where filtered ischemes would still be used
This was caused by not having a proper index for the mnemonic in the
insnfeaturemanager, but rather using the mapping from the iwho context,
which ignores filters.
As a result, among other things, sampling in all previous runs would
have been somewhat broken.

The fix is to just remove all special handling for the mnemonic feature
and have it handled like all others.